方法重构
