+++ SwiftExpand.podspec
@@ -2,7 +2,7 @@
 Pod::Spec.new do |s|
 
     s.name         = "SwiftExpand"
-    s.version      = "7.2.0.4"
+    s.version      = "7.2.0.5"
     s.summary      = "系统类功能扩展, 极大的提高工作效率."
     s.description  = "主要通过分类实现"
